Make benchmarks a little easier to work with
