+++ bench/benchmarks/buffer.js
@@ -4,14 +4,14 @@ var VT = require('vector-tile');
 var Protobuf = require('pbf');
 var assert = require('assert');
 
-var WorkerTile = require('../js/source/worker_tile');
-var ajax = require('../js/util/ajax');
-var Style = require('../js/style/style');
-var util = require('../js/util/util');
-var Evented = require('../js/util/evented');
-var config = require('../js/util/config');
-var coordinates = require('./coordinates');
-var formatNumber = require('./format_number');
+var WorkerTile = require('../../js/source/worker_tile');
+var ajax = require('../../js/util/ajax');
+var Style = require('../../js/style/style');
+var util = require('../../js/util/util');
+var Evented = require('../../js/util/evented');
+var config = require('../../js/util/config');
+var coordinates = require('../coordinates');
+var formatNumber = require('../format_number');
 
 var SAMPLE_COUNT = 10;
