+++ bench/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-    <title>Mapbox GL JS Buffer Benchmark</title>
+    <title>Mapbox GL JS Benchmarks</title>
 
     <meta charset='utf-8'>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -10,6 +10,14 @@
     <link rel='stylesheet' href='/dist/mapbox-gl.css'/>
 
     <style>
+        #benchmarks {
+            width: 100%;
+            margin: 25px auto;
+            text-align: center;
+        }
+        #benchmarks a {
+            margin: 5px;
+        }
         #logs {
             margin: 25px auto;
             width: 600px;
@@ -25,6 +33,7 @@
 </head>
 
 <body class="prose">
+    <div id="benchmarks"></div>
     <div id="logs"></div>
     <div id="map"></div>
     <script src="/bench/index.js"></script>
