+++ bench/index.js
@@ -14,11 +14,17 @@ function main() {
     log('dark', 'please keep this window in the foreground and close the debugger');
 
     var benchmarks = {
-        buffer: require('./buffer_benchmark'),
-        fps: require('./fps_benchmark'),
-        'query-point': require('./query_point_benchmark'),
-        'query-box': require('./query_box_benchmark')
+        buffer: require('./benchmarks/buffer'),
+        fps: require('./benchmarks/fps'),
+        'query-point': require('./benchmarks/query_point'),
+        'query-box': require('./benchmarks/query_box')
     };
+
+    var benchmarksDiv = document.getElementById('benchmarks');
+
+    Object.keys(benchmarks).forEach(function(id) {
+        benchmarksDiv.innerHTML += '<a href="/bench/' + id + '" class="button">' + id + '</a>';
+    });
 
     var pathnameArray = location.pathname.split('/');
     var benchmarkName = pathnameArray[pathnameArray.length - 1] || pathnameArray[pathnameArray.length - 2];
